Move RestScreen to its own module
