+++ tests/test_ui.py
@@ -19,7 +19,6 @@ if kivy_available:
     import core
 
     from main import (
-        RestScreen,
         MetricInputScreen,
         WorkoutActiveScreen,
         AddMetricPopup,
@@ -30,6 +29,7 @@ if kivy_available:
         PresetsScreen,
         EditPresetScreen,
     )
+    from ui.screens.rest_screen import RestScreen
     import time
 
     class _DummyApp:
